requirements.txt
# Data manipulation
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Machine Learning
scikit-learn>=1.3.0
prophet>=1.1.4  # Time series forecasting with seasonality

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0

# Dashboard
streamlit>=1.28.0

# Utilities
holidays>=0.35
python-dateutil>=2.8.0
joblib>=1.3.0

# Security
bcrypt>=4.0.0
python-dotenv>=1.0.0

# Excel export
xlsxwriter>=3.1.0

# API (optional)
flask>=3.0.0

# Statistics
scipy>=1.11.0
statsmodels>=0.14.0

